feat(relayer): Work with multiple networks

+++ beacon-light-client/circom/scripts/light_client/workers/proof-generator-worker.ts
@@ -1,4 +1,4 @@
-import { Worker, Queue } from 'bullmq';
+import { Worker } from 'bullmq';
 import { exec as _exec } from 'child_process';
 import { readFile, rm, writeFile } from 'fs/promises';
 import path from 'path';
@@ -9,23 +9,28 @@ import {
   PROOF_GENERATOR_QUEUE,
   RELAYER_INPUTS_FOLDER,
   RELAYER_PROOFS_FOLDER,
-  PUBLISH_ONCHAIN_QUEUE,
-  ProofResultType,
+  getProofKey,
+  PROOFS_CHANEL,
 } from '../relayer-helper';
 import * as config from '../config.json';
-
+import redisClient from '../client';
 const exec = promisify(_exec);
 
-const proofPublishQueue = new Queue<ProofResultType>(PUBLISH_ONCHAIN_QUEUE, {
-  connection: {
-    host: config.redisHost,
-    port: config.redisPort,
-  },
-});
+const publisher = redisClient.duplicate();
 
 new Worker<ProofInputType>(
   PROOF_GENERATOR_QUEUE,
   async job => {
+    const existingProof = await redisClient.get(
+      `proof:${job.data.prevUpdateSlot}:${job.data.updateSlot}`,
+    );
+
+    if (existingProof !== null) {
+      await publisher.publish(PROOFS_CHANEL, 'proof');
+
+      return;
+    }
+
     await writeFile(
       path.join(
         __dirname,
@@ -103,17 +108,26 @@ new Worker<ProofInputType>(
       ),
     );
 
-    await proofPublishQueue.add('proofGenerate', {
-      prevUpdateSlot: job.data.prevUpdateSlot,
-      updateSlot: job.data.updateSlot,
-      proofInput: job.data.proofInput,
-      proof: {
-        pi_a: proof.pi_a,
-        pi_b: proof.pi_b,
-        pi_c: proof.pi_c,
-        public: publicVars,
-      },
-    });
+    await redisClient.set(
+      getProofKey(job.data.prevUpdateSlot, job.data.updateSlot),
+      JSON.stringify({
+        prevUpdateSlot: job.data.prevUpdateSlot,
+        updateSlot: job.data.updateSlot,
+        proofInput: job.data.proofInput,
+        proof: {
+          pi_a: proof.pi_a,
+          pi_b: proof.pi_b,
+          pi_c: proof.pi_c,
+          public: publicVars,
+        },
+      }),
+    );
+
+    if (!publisher.isOpen) {
+      await publisher.connect();
+    }
+
+    await publisher.publish(PROOFS_CHANEL, 'proof');
   },
   {
     connection: {
